+++ CS350_BaggingApplication/Models/Packaging.cs
@@ -4,10 +4,10 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int WeightCapacity{ get; set; }
+        public double WeightCapacity{ get; set; }
         public int HardItemLimit { get; set; }
-        public int Height { get; set; }
-        public int Width { get; set; }
-        public int Length { get; set; }
+        public double Height { get; set; }
+        public double Width { get; set; }
+        public double Length { get; set; }
     }
 }
